+++ src/main/java/net/rasanovum/rockandstone/RockAndStoneConfig.java
@@ -4,4 +4,6 @@ import eu.midnightdust.lib.config.MidnightConfig;
 
 public class RockAndStoneConfig extends MidnightConfig {
     @Entry public static boolean doRockAndStoneDebug = false;
+    @Entry public static boolean doOreReplacement = true;
+    @Entry public static boolean doSilentAdvancements = false;
 }